renaming controller and moving it to a seperate file
