+++ gallery_ui/fancybox/code/FancyBox.php
@@ -11,7 +11,7 @@ class FancyBox extends ImageGalleryUI {
 	public $item_template = "FancyBox_item";
 
 	public function initialize() {
-		Requirements::javascript(THIRDPARTY_DIR . '/jquery/jquery.js');
+		Requirements::javascript('silverstripe-admin/thirdparty/jquery/jquery.js');
 		Requirements::javascript('image_gallery/gallery_ui/fancybox/javascript/jquery.fancybox.js');
 		Requirements::javascript('image_gallery/gallery_ui/fancybox/javascript/jquery.pngFix.pack.js');
 		Requirements::javascript('image_gallery/gallery_ui/fancybox/javascript/fancybox_init.js');
