+++ gallery_ui/lightbox/code/Lightbox.php
@@ -1,4 +1,4 @@
-<?php 
+<?php
 use SilverStripe\View\Requirements;
 
 
@@ -7,14 +7,14 @@ class Lightbox extends ImageGalleryUI
 	static $link_to_demo = "http://leandrovieira.com/projects/jquery/lightbox/";
 	static $label = "LightBox";
 	public $item_template = "Lightbox_item";
-	
+
 	public function initialize()
 	{
-		Requirements::javascript(THIRDPARTY_DIR.'/jquery/jquery.js'); 
+        Requirements::javascript('silverstripe-admin/thirdparty/jquery/jquery.js');
 		Requirements::javascript('image_gallery/gallery_ui/lightbox/javascript/jquery.lightbox-0.5.js');
 		Requirements::javascript('image_gallery/gallery_ui/lightbox/javascript/lightbox_init.js');
 		Requirements::css('image_gallery/gallery_ui/lightbox/css/jquery.lightbox-0.5.css');
-		
+
 	}
-	
+
 }
